[bucket C] revert conftest.py: restore self-contained per-file test fixtures
conftest.py was the only one in the entire repo -- every other test
module here duplicates its own small fixtures rather than sharing them
across files. Reverts to that convention: each of the three consumers
gets its own local _CustomModelFixture and setUp again.
